+++ Sources/Core/Foundation/CurrentLocaleMonitor.swift
@@ -11,15 +11,15 @@ import Foundation
 
 ///
 /// A `CurrentLocaleMonitor` instance monitors the system for changes to the
-/// user's locale.
+/// user’s locale.
 ///
 public class CurrentLocaleMonitor: BaseNotificationMonitor {
     ///
-    /// Encapsulates changes to the user's locale.
+    /// Encapsulates changes to the user’s locale.
     ///
     public enum Event {
         ///
-        /// The user's locale did change.
+        /// The user’s locale has changed.
         ///
         case didChange
     }
@@ -30,11 +30,12 @@ public class CurrentLocaleMonitor: BaseNotificationMonitor {
     /// - Parameters:
     ///   - queue:      The operation queue on which the handler executes.
     ///                 By default, the main operation queue is used.
-    ///   - handler:    The handler to call when the user's locale did change.
+    ///   - handler:    The handler to call when the user’s locale changes.
     ///
     public init(queue: OperationQueue,
                 handler: @escaping (Event) -> Void) {
         self.handler = handler
+
         super.init(queue: queue)
     }
 
@@ -42,6 +43,7 @@ public class CurrentLocaleMonitor: BaseNotificationMonitor {
 
     public override func addNotificationObservers() {
         super.addNotificationObservers()
+
         observe(NSLocale.currentLocaleDidChangeNotification) { [unowned self] _ in
             self.handler(.didChange)
         }