+++ Sources/Core/Foundation/URLCredentialStorageMonitor.swift
@@ -10,14 +10,17 @@
 import Foundation
 
 ///
-/// A `URLCredentialStorageMonitor` instance monitors the system for changes
-/// in the credentials storage
+/// A `URLCredentialStorageMonitor` instance monitors the shared URL credential
+/// storage object for changes to its stored credentials.
 ///
 public class URLCredentialStorageMonitor: BaseNotificationMonitor {
     ///
-    /// Encapsulates changes to the url credentials storage.
+    /// Encapsulates changes to the credential storage.
     ///
     public enum Event {
+        ///
+        /// The stored credentials have changed.
+        ///
         case changed(URLCredentialStorage)
     }
 
@@ -27,33 +30,27 @@ public class URLCredentialStorageMonitor: BaseNotificationMonitor {
     /// - Parameters:
     ///   - queue:      The operation queue on which the handler executes.
     ///                 By default, the main operation queue is used.
-    ///   - handler:    The handler to call when the bundle dynamically loads
-    ///                 classes.
+    ///   - handler:    The handler to call when the set of stored URL
+    ///                 credentials changes.
     ///
     public init(queue: OperationQueue = .main,
                 handler: @escaping (Event) -> Void) {
+        self.credentialStorage = .shared
         self.handler = handler
 
         super.init(queue: queue)
     }
 
+    private let credentialStorage: URLCredentialStorage
     private let handler: (Event) -> Void
-
-    private func extractUrlCredentialStorage(_ notification: Notification) -> URLCredentialStorage? {
-        guard
-            let urlCredentialStorage = notification.object as? URLCredentialStorage
-            else { return nil }
-
-        return urlCredentialStorage
-    }
 
     override public func  addNotificationObservers() {
         super.addNotificationObservers()
 
         observe(.NSURLCredentialStorageChanged,
-                object: URLCredentialStorage.shared) { [unowned self] in
-                    if let urlCredentialStorage = self.extractUrlCredentialStorage($0) {
-                        self.handler(.changed(urlCredentialStorage))
+                object: credentialStorage) { [unowned self] in
+                    if let credentialStorage = $0.object as? URLCredentialStorage {
+                        self.handler(.changed(credentialStorage))
                     }
         }
     }