+++ Tests/UIKit/Other/MenuControllerMonitorTests.swift
@@ -18,9 +18,7 @@ internal class MenuControllerMonitorTests: XCTestCase {
     override func setUp() {
         super.setUp()
 
-        NotificationCenterInjector.inject = {
-            return self.notificationCenter
-        }
+        NotificationCenterInjector.inject = { return self.notificationCenter }
     }
 
     func testMonitor_didHideMenu() {